Provide a working implementation of the User access token
At the moment, the template ships with the Shop access token resource
only.

To add User access token capabilities, devs need to run:

rails generate shopify_app:user_model

run the auto-generated migrations, and change a setting in
config/initializers/shopify_app.rb, as described in the
shopify_app gem documentation [1].

With this change, we are providing all the code to have the application
working with both approaches without running any extra generator and
migration. The only thing left to do for devs, is to uncomment the setting,
which has been better documented as well.

As a positive side-effect, this change also solves an inconsistency of
the current codebase, which was including the user table in the database
schema, without having its migrations.

[1]: https://github.com/Shopify/shopify_app/blob/b921685a0dbf1948905cd3057d35d2f484cdbce8/docs/shopify_app/sessions.md#user-online-token-storage